+++ sql/postgres/driver_test.go
@@ -3,11 +3,9 @@ package postgres_test
 import (
 	"context"
 	"database/sql"
-	"os"
 
 	"github.com/dogmatiq/projectionkit/sql/internal/drivertest"
 	. "github.com/dogmatiq/projectionkit/sql/postgres"
-	_ "github.com/lib/pq"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
@@ -16,14 +14,7 @@ var _ = Describe("type Driver", func() {
 	var db *sql.DB
 
 	BeforeSuite(func() {
-		dsn := os.Getenv("DOGMATIQ_TEST_POSTGRES_DSN")
-		if dsn == "" {
-			dsn = "user=postgres password=rootpass sslmode=disable"
-		}
-
-		var err error
-		db, err = sql.Open("postgres", dsn)
-		Expect(err).ShouldNot(HaveOccurred())
+		db = drivertest.Open("postgres")
 	})
 
 	AfterSuite(func() {