+++ sql/sqlite/driver_test.go
@@ -3,12 +3,10 @@ package sqlite_test
 import (
 	"context"
 	"database/sql"
-	"os"
 	"strings"
 
 	"github.com/dogmatiq/projectionkit/sql/internal/drivertest"
 	. "github.com/dogmatiq/projectionkit/sql/sqlite"
-	_ "github.com/mattn/go-sqlite3"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
@@ -17,18 +15,7 @@ var _ = Describe("type Driver", func() {
 	var db *sql.DB
 
 	BeforeSuite(func() {
-		dsn := os.Getenv("DOGMATIQ_TEST_SQLITE_DSN")
-		if dsn == "" {
-			dsn = ":memory:"
-		}
-
-		var err error
-		db, err = sql.Open("sqlite3", dsn)
-		Expect(err).ShouldNot(HaveOccurred())
-
-		// Ensure that we only ever have one "connection" to the memory
-		// database, otherwise the only be created in one of them.
-		db.SetMaxOpenConns(1)
+		db = drivertest.Open("sqlite3")
 	})
 
 	AfterSuite(func() {